# api-docs/openapi.yaml
openapi: 3.1.0
info:
  title: Multi-Tenant IAM API
  version: 1.0.0
  description: API para gestión de identidad y acceso multi-tenant
servers:
  - url: /iam/api/v1
    description: Base URL para la API v1
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT para autenticación
    tenantId:
      type: apiKey
      in: header
      name: X-Tenant-ID
      description: ID del tenant para operaciones multi-tenant
...
  schemas:
...
    ListResponse:
      type: object
      properties:
        items:
          type: array
          items: {}
        total_count:
          type: integer
          description: Total de elementos
        page:
          type: integer
          description: Página actual
        page_size:
          type: integer
          description: Tamaño de página
        total_pages:
          type: integer
          description: Total de páginas

    # Respuestas específicas usando ListResponse
...
    TenantListResponse:
      allOf:
        - $ref: '#/components/schemas/ListResponse'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Tenant'
...
    Tenant:
      type: object
      required:
        - name
        - slug
        - type
        - owner_id
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 2
          maxLength: 100
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
          minLength: 2
          maxLength: 50
        description:
          type: string
          minLength: 5
          maxLength: 500
        type:
          type: string
          enum: [PERSONAL, STARTUP, BUSINESS, ENTERPRISE]
        status:
          type: string
          enum: [ACTIVE, INACTIVE, SUSPENDED, DELETED]
        owner_id:
          type: string
          format: uuid
        domain:
          type: string
          format: hostname
        plan_id:
          type: string
          format: uuid
          nullable: true
        user_count:
          type: integer
          minimum: 0
        max_users:
          type: integer
          minimum: -1
        settings:
          type: object
          additionalProperties: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        is_active:
          type: boolean
          readOnly: true
        can_add_users:
          type: boolean
          readOnly: true
        has_custom_domain:
          type: boolean
          readOnly: true
    CreateTenantRequest:
      type: object
      required:
        - name
        - slug
        - type
        - owner_id
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
          minLength: 2
          maxLength: 50
        description:
          type: string
          minLength: 5
          maxLength: 500
        type:
          type: string
          enum: [PERSONAL, STARTUP, BUSINESS, ENTERPRISE]
        owner_id:
          type: string
          format: uuid
        domain:
          type: string
          format: hostname
    UpdateTenantRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        description:
          type: string
          minLength: 5
          maxLength: 500
        domain:
          type: string
          format: hostname
        status:
          type: string
          enum: [ACTIVE, INACTIVE, SUSPENDED]
        type:
          type: string
          enum: [PERSONAL, STARTUP, BUSINESS, ENTERPRISE]
        settings:
          type: object
          additionalProperties: true
    SetPlanRequest:
      type: object
      required:
        - plan_id
      properties:
        plan_id:
          type: string
          format: uuid
...
paths:
...
  /tenants:
    post:
      summary: Crear nuevo tenant
      tags: [Tenants]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
      responses:
        '201':
          description: Tenant creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '400':
          description: Datos de entrada inválidos
        '409':
          description: Slug o dominio ya existe

    get:
      summary: Listar tenants con filtros
      tags: [Tenants]
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: owner_id
          in: query
          description: Filtrar por propietario
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          description: Filtrar por estado
          schema:
            type: string
            enum: [ACTIVE, INACTIVE, SUSPENDED, DELETED]
        - name: type
          in: query
          description: Filtrar por tipo
          schema:
            type: string
            enum: [PERSONAL, STARTUP, BUSINESS, ENTERPRISE]
        - name: active
          in: query
          description: Solo tenants activos
          schema:
            type: boolean
        - name: expiring_days
          in: query
          description: Tenants que expiran en X días
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Lista de tenants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantListResponse'
  /tenants/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Obtener tenant por ID
      tags: [Tenants]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Tenant encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '404':
          description: Tenant no encontrado

    put:
      summary: Actualizar tenant
      tags: [Tenants]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTenantRequest'
      responses:
        '200':
          description: Tenant actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '400':
          description: Datos de entrada inválidos
        '403':
          description: No se puede modificar tenant eliminado
        '404':
          description: Tenant no encontrado
        '409':
          description: Dominio ya existe

    delete:
      summary: Eliminar tenant (soft delete)
      tags: [Tenants]
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Tenant eliminado
        '403':
          description: No se puede eliminar el tenant
        '404':
          description: Tenant no encontrado
  /tenants/by-slug/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          pattern: '^[a-z0-9-]+$'
    get:
      summary: Obtener tenant por slug
      tags: [Tenants]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Tenant encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '404':
          description: Tenant no encontrado
  /tenants/{id}/plan:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Asignar plan a tenant
      tags: [Tenants]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetPlanRequest'
      responses:
        '200':
          description: Plan asignado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '400':
          description: Plan no válido
        '403':
          description: Tenant no activo o eliminado
        '404':
          description: Tenant no encontrado

    delete:
      summary: Remover plan del tenant
      tags: [Tenants]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Plan removido exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '403':
          description: Tenant eliminado
        '404':
          description: Tenant no encontrado
...